controllers

+++ src/main/java/com/example/demo/domain/services/interfaces/EventoDomainService.java
@@ -23,5 +23,5 @@ public interface EventoDomainService {
 	
 	List<EventoResponseDto> consultarEventosPorStatus(String status);
 	
-	long contarEventos();
+	Integer contarEventos();
 }
